move user controller to new format
