+++ api/src/controllers/index.ts
@@ -9,4 +9,4 @@ export * from "./conversation.controller/";
 export * from "./inbound.controller/";
 export * from "./integration.controller/";
 export * from "./team.controller/";
-export * from "./user.controller";
+export * from "./user.controller/";
